Refactor: Consolidate DI EntryPoints into logical domain groups
- Removed overly granular entry points (e.g. WorkerDependenciesEntryPoint, AutoSyncEntryPoint, ServiceEntryPoint).
- Created logically grouped domain entry points (CoreDependenciesEntryPoint, NetworkDependenciesEntryPoint, RepositoryDependenciesEntryPoint, ServiceDependenciesEntryPoint).
- Migrated all usages of EntryPointAccessors across MainApplication, utilities, and tests to use the new domain entry points.
- Addressed `must-fix` compiler feedback by completely deleting the old interfaces rather than annotating them with `@Deprecated`.
- Mitigated "service locator" anti-pattern by splitting dependencies into functional domains instead of a single mega-interface.
