+++ app/src/main/java/org/ole/planet/myplanet/MainApplication.kt
@@ -34,12 +34,9 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.callback.OnTeamPageListener
 import org.ole.planet.myplanet.data.DatabaseService
-import org.ole.planet.myplanet.di.ApiClientEntryPoint
-import org.ole.planet.myplanet.di.ApplicationScopeEntryPoint
+import org.ole.planet.myplanet.di.NetworkDependenciesEntryPoint
+import org.ole.planet.myplanet.di.CoreDependenciesEntryPoint
 import org.ole.planet.myplanet.di.DefaultPreferences
-import org.ole.planet.myplanet.di.RetryQueueEntryPoint
-import org.ole.planet.myplanet.di.ServerUrlMapperEntryPoint
-import org.ole.planet.myplanet.di.SharedInternalEntryPoint
 import org.ole.planet.myplanet.model.RealmApkLog
 import org.ole.planet.myplanet.repository.ResourcesRepository
 import org.ole.planet.myplanet.services.AutoSyncWorker
@@ -115,10 +112,10 @@ class MainApplication : Application(), Application.ActivityLifecycleCallbacks, W
             applicationScope.launch {
                 val entryPoint = EntryPointAccessors.fromApplication(
                     context,
-                    SharedInternalEntryPoint::class.java
+                    CoreDependenciesEntryPoint::class.java
                 )
                 val userSessionManager = entryPoint.userSessionManager()
-                val spm = EntryPointAccessors.fromApplication(context, SharedInternalEntryPoint::class.java).sharedPrefManager()
+                val spm = EntryPointAccessors.fromApplication(context, CoreDependenciesEntryPoint::class.java).sharedPrefManager()
                 try {
                     val databaseService = (context.applicationContext as MainApplication).databaseService
                     val model = userSessionManager.getUserModel()
@@ -153,7 +150,7 @@ class MainApplication : Application(), Application.ActivityLifecycleCallbacks, W
             urlString: String,
             ioDispatcher: kotlinx.coroutines.CoroutineDispatcher = kotlinx.coroutines.Dispatchers.IO
         ): Boolean {
-            val entryPoint = EntryPointAccessors.fromApplication(context, ServerUrlMapperEntryPoint::class.java)
+            val entryPoint = EntryPointAccessors.fromApplication(context, CoreDependenciesEntryPoint::class.java)
             val serverUrlMapper = entryPoint.serverUrlMapper()
             val mapping = serverUrlMapper.processUrl(urlString)
             val urlsToTry = mutableListOf(urlString)
@@ -238,7 +235,7 @@ class MainApplication : Application(), Application.ActivityLifecycleCallbacks, W
     private fun setupCriticalProperties() {
         applicationScope = EntryPointAccessors.fromApplication(
             this,
-            ApplicationScopeEntryPoint::class.java
+            CoreDependenciesEntryPoint::class.java
         ).applicationScope()
     }
 
@@ -246,7 +243,7 @@ class MainApplication : Application(), Application.ActivityLifecycleCallbacks, W
         withContext(dispatcherProvider.io) {
             EntryPointAccessors.fromApplication(
                 this@MainApplication,
-                ApiClientEntryPoint::class.java
+                NetworkDependenciesEntryPoint::class.java
             ).apiClient()
             apiClientInitialized.complete(Unit)
         }
@@ -305,7 +302,7 @@ class MainApplication : Application(), Application.ActivityLifecycleCallbacks, W
             try {
                 val entryPoint = EntryPointAccessors.fromApplication(
                     this@MainApplication,
-                    RetryQueueEntryPoint::class.java
+                    NetworkDependenciesEntryPoint::class.java
                 )
                 entryPoint.retryQueue().recoverStuckOperations()
             } catch (e: Exception) {
