--- app/src/main/java/org/ole/planet/myplanet/di/ApiClientEntryPoint.kt
@@ -1,13 +1,0 @@
-package org.ole.planet.myplanet.di
-
-import dagger.hilt.EntryPoint
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import org.ole.planet.myplanet.data.api.ApiClient
-
-@Deprecated("Use SharedInternalEntryPoint instead")
-@EntryPoint
-@InstallIn(SingletonComponent::class)
-interface ApiClientEntryPoint {
-    fun apiClient(): ApiClient
-}
