--- app/src/main/java/org/ole/planet/myplanet/di/ApiInterfaceEntryPoint.kt
@@ -1,13 +1,0 @@
-package org.ole.planet.myplanet.di
-
-import dagger.hilt.EntryPoint
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import org.ole.planet.myplanet.data.api.ApiInterface
-
-@Deprecated("Use SharedInternalEntryPoint instead")
-@EntryPoint
-@InstallIn(SingletonComponent::class)
-interface ApiInterfaceEntryPoint {
-    fun apiInterface(): ApiInterface
-}
