--- app/src/main/java/org/ole/planet/myplanet/di/ApplicationScopeEntryPoint.kt
@@ -1,14 +1,0 @@
-package org.ole.planet.myplanet.di
-
-import dagger.hilt.EntryPoint
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineScope
-
-@Deprecated("Use SharedInternalEntryPoint instead")
-@EntryPoint
-@InstallIn(SingletonComponent::class)
-interface ApplicationScopeEntryPoint {
-    @ApplicationScope
-    fun applicationScope(): CoroutineScope
-}
