--- app/src/main/java/org/ole/planet/myplanet/di/DatabaseServiceEntryPoint.kt
@@ -1,13 +1,0 @@
-package org.ole.planet.myplanet.di
-
-import dagger.hilt.EntryPoint
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import org.ole.planet.myplanet.data.DatabaseService
-
-@Deprecated("Use SharedInternalEntryPoint instead")
-@EntryPoint
-@InstallIn(SingletonComponent::class)
-interface DatabaseServiceEntryPoint {
-    fun databaseService(): DatabaseService
-}
