+++ app/src/main/java/org/ole/planet/myplanet/di/NetworkDependenciesEntryPoint.kt
@@ -3,17 +3,20 @@ package org.ole.planet.myplanet.di
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import org.ole.planet.myplanet.services.SharedPrefManager
+import org.ole.planet.myplanet.data.api.ApiClient
+import org.ole.planet.myplanet.data.api.ApiInterface
 import org.ole.planet.myplanet.services.UploadManager
 import org.ole.planet.myplanet.services.UploadToShelfService
+import org.ole.planet.myplanet.services.retry.RetryQueue
 import org.ole.planet.myplanet.services.sync.SyncManager
 
-@Deprecated("Use SharedInternalEntryPoint instead")
 @EntryPoint
 @InstallIn(SingletonComponent::class)
-interface AutoSyncEntryPoint {
+interface NetworkDependenciesEntryPoint {
+    fun apiClient(): ApiClient
+    fun apiInterface(): ApiInterface
     fun syncManager(): SyncManager
     fun uploadManager(): UploadManager
     fun uploadToShelfService(): UploadToShelfService
-    fun sharedPrefManager(): SharedPrefManager
+    fun retryQueue(): RetryQueue
 }
