+++ app/src/main/java/org/ole/planet/myplanet/di/RepositoryDependenciesEntryPoint.kt
@@ -5,13 +5,18 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.ole.planet.myplanet.repository.CommunityRepository
 import org.ole.planet.myplanet.repository.ConfigurationsRepository
+import org.ole.planet.myplanet.repository.ResourcesRepository
+import org.ole.planet.myplanet.repository.SubmissionsRepository
+import org.ole.planet.myplanet.repository.TeamsRepository
 import org.ole.planet.myplanet.repository.UserRepository
 
-@Deprecated("Use SharedInternalEntryPoint instead")
 @EntryPoint
 @InstallIn(SingletonComponent::class)
-interface RepositoryEntryPoint {
+interface RepositoryDependenciesEntryPoint {
     fun userRepository(): UserRepository
     fun configurationsRepository(): ConfigurationsRepository
     fun communityRepository(): CommunityRepository
+    fun teamsRepository(): TeamsRepository
+    fun submissionsRepository(): SubmissionsRepository
+    fun resourcesRepository(): ResourcesRepository
 }
