+++ fix_reviewer_bug5.py
@@ -1,0 +1,3 @@
+# Ah!
+# Let me search for `EntryPointAccessors` in ANY file, regardless of extension.
+# Maybe `WorkerDependenciesEntryPoint` is used somewhere I missed.
